ConnectDialog: disable ping and host lookups when using a proxy.
This commit adds two flags to the ConnectDialog. One that allows
pings, and one that allows hostname lookups.

By default, they are both enabled.

However, if you're using a proxy, both will be disabled.

The commit also updates the ConnectDialog UI to hide
the user count and ping columns when pings are not enabled.

As a consequence of this commit, Mumble will now allow you to
connect to a server for which a hostname has not yet been
resolved if you are using a proxy.

Previously, for hosts that could not be resolved via the regular
resolvers in Mumble (system resolver, Bonjour, ...), Mumble
would not be able to connect. This is because Mumble would disable
the Connect button if no hostname(s) could be resolved -- or were
not yet resolved.

With this commit, Mumble will now allow connections to those
hosts, andæ let the proxy will do the hostname resolving. This
allows, among other things, Mumble to connect to Tor .onion
hosts when using Tor as a proxy.

Fixes mumble-voip/mumble#1033
Fixes mumble-voip/mumble#1812