Murmur: allow both bitmasked version and version string when setting 'suggestversion' via RPC.
This commit ensures that a 'suggestversion' is always normalized to a
bitmasked version.

Previously, only a bitmasked 'suggestverson' was allowed.

With this commit, 'suggestversion' may now be either a version string,
or a bitmasked integer value representing the version.

The previous behavior was confusing to users, since the murmur.ini
has always used 'suggestversion' with a version string.

Fixes mumble-voip/mumble#989