GlobalShortcut_win, mumble.pro: remove use of HardHook in GlobalShortcut_win.
This also lets us remove MinHook from the Mumble client itself.

This code is part of the client-in-overlay feature that we had
for Windows and OS X. However, that feature is broken for Qt 5
and never had good performance.

The reason I am removing this code is that MinHook is built against
the static C runtime. As such, we get warnings from the linker
when linking in MinHook, because the Mumble client expects to be
linked against the MultiThreadedDLL runtime.