Simplify LICENSE file, add LICENSE.header and add AUTHORS.
This simplifies the LICENSE file refer to the copyright
holders of Mumble as "The Mumble Developers". The client
already does this in the About dialog.

The entity "The Mumble Developers" is the name we use for
the copyright holders of Mumble. These are listed in the
AUTHORS file.

The AUTHORS file is generated via scripts/generate-AUTHORS.py.
The script looks at the Git history and removes duplicates and
other mistakes made through the years.

All new files in the repo should use the license header found
in LICENSE.header.

The AUTHORS and LICENSE files are permalinked to

    https://www.mumble.info/LICENSE
    https://www.mumble.info/AUTHORS

These locations are used in the files themselves,
as well as the license header.