Add appveyor configuration for windows proof builds
The current configuration uses a hacked up
win64-static-no-ltcg build environment until we
build our next set of real ones. As we do not want
to distribute our build env at this point the source
is obfuscated for now.

We utilize the appveyor cache to cache the compressed
build environment we downloaded. We do not cache the
uncompressed variant to save the time it takes to package
it up in case of cache invalidation. Cache invalidation
is keyed on changes to the appveyor.yml .