installer: use WiX from buildenv.
This uses the mechanism documented at
http://wixtoolset.org/documentation/manual/v3/msbuild/daily_builds.html
to use a local copy of WiX instead of using the WiX MSI package.

Previously, we required people to install WiX globally.
That's yet another tedious dependency required in order to build Mumble.

We'll now be distributing WiX in our buildenvs, so tell our .wixproj
to use that version.