Meta: avoid cluttering the global QSslSocket::defaultCaCertificates() list.
Before this commit, we would add any intermediate certificates found in
the SSL certificate's PEM bundle to the default CA list.
We would also add any certificates found in the sslCA PEM bundle to the
default CA list.

Due to the way the hot-reload feature for SSL settings is going to work,
these patterns wouldn't work well in that environment, because reloading
the settings would keep adding entries to the global QSslSocket default
CA certificates list.

To combat this, this series of commits try to be more strict about how
default CAs and intermedaites in general are handled.
Generally, instead of adding them to the global default CA list, we add
them to each QSslSocket as necessary.