CryptographicHash: new class for computing cryptographic hashes.
We used to use QCryptographicHash, but Qt 4 doesn't support Sha256.

To remedy that, we've implemented this class on top of OpenSSL's EVP
system.

Besides being able to work on Qt 4, this class also means we can add
additional hash functions in the future, which are not necessarily
supported by Qt, such as BLAKE2.