Meta: move qmConfig.clear().
This moves the qmConfig.clear() call in MetaParams::read() to the
top of the function.

This moves the qmConfig.clear() option to the top of MetaParams::read()
instead.

This allows the SSL part of MetaParams::read() to be moved out to a
separate function in the hot certificate reload PR.