OverlayConfig: do not show built-in whitelist items when in WhitelistExclusionMode.
To support the launcher filter overlay mode, we've added a default set of
programs to Mumble's whitelist.

However, when in 'whitelist' mode, Mumble has traditionally not provided
any defaults.

Also, for the whitelist mode to be useful to people, it should not have
any defaults -- because the whole purpose of the whitelist is for people
to be 100% in control themselves.

This commit changes the UI in OverlayConfig to not show the defaults
from overlay_whitelist.h when in WhitelistExclusionMode.