OverlayConfig: add UI for configuring the launcher filter.
This change adds the user interface and config dialog logic
needed for the overlay launcher filter.

This adds an additional mode to the overlay called the 'launcher filter'.

The idea behind the launcher filter is that we provide a sensible
whitelist of allowed launchers (which are typically digital storefronts,
and other launchers -- Steam, Origin, GOG, Battle.net, MMORPG launchers,
etc.), as well as a default whitelist of normal programs, because some
things don't fit into the 'launcher' check -- such as Guild Wars 2.

The existing 'blacklist' and 'whitelist' modes are still available,
as before.

This commit also changes the way "default" entries are shown.
Instead of being in a disabled state, they are now shown in italic.
Also, default entries now have a checkbox next to them, allowing users
to disable default entries.