scripts/mklic.pl: add MariaDB Connector/C license.
This is shipped with all our official binary copies of Murmur
(except on macOS at the moment, but that's a historic accident...)

This is just the short form of LGPL 2.1. That's all they ship in
COPYING.LIB.

Since the MariaDB Connector/C library is based on an old libmysql,
the copyright is all over the place, which I guess is why they don't
have the copyright notices anywhere. (It's messy, I tried to compile
a list and gave up.)